Small improvements in drift metadata export

+++ scripts/proc/export_polish_metadata.com
@@ -31,7 +31,7 @@ if ( ${sub_doit} == "y" ) then
 
         endif 
 
-        if ( ! -e motioncor2_shifts.star ) then
+        if ( -e LOGS/motioncor2.out ) then
 
           set imsize_x = `grep "Stack size" LOGS/motioncor2.out | head -n 1 | cut -d' ' -f3`
           set imsize_y = `grep "Stack size" LOGS/motioncor2.out | head -n 1 | cut -d' ' -f5`
@@ -39,40 +39,47 @@ if ( ${sub_doit} == "y" ) then
           cat > ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star << eot
 
 
-            data_general
+data_general
 
-            _rlnImageSizeX                                     ${imsize_x}
-            _rlnImageSizeY                                     ${imsize_y}
-            _rlnImageSizeZ                                     ${movie_imagenumber_total}
-            _rlnMicrographMovieName                            ${export_rawstack_subdir}/${import_rawstack}
-            _rlnMicrographGainName                             ${export_gainref_subdir}/${gainref_name:r}.mrc
-            _rlnMicrographBinning                              ${bin_factor}
-            _rlnMicrographOriginalPixelSize                    ${pixelsize}
-            _rlnMicrographDoseRate                             ${frame_dose}
-            _rlnMicrographPreExposure                          0.000000
-            _rlnVoltage                                        ${KV}
-            _rlnMicrographStartFrame                           1
-            _rlnMotionModelVersion                             1
-             
+_rlnImageSizeX                                     ${imsize_x}
+_rlnImageSizeY                                     ${imsize_y}
+_rlnImageSizeZ                                     ${movie_imagenumber_total}
+_rlnMicrographMovieName                            ${export_rawstack_subdir}/${import_rawstack}
+_rlnMicrographGainName                             ${export_gainref_subdir}/${gainref_name:r}.mrc
+_rlnMicrographBinning                              ${bin_factor}
+_rlnMicrographOriginalPixelSize                    ${pixelsize}
+_rlnMicrographDoseRate                             ${frame_dose}
+_rlnMicrographPreExposure                          0.000000
+_rlnVoltage                                        ${KV}
+_rlnMicrographStartFrame                           1
+_rlnMotionModelVersion                             1
+ 
 
-            data_global_shift
+data_global_shift
 
-            loop_ 
-            _rlnMicrographFrameNumber #1 
-            _rlnMicrographShiftX #2 
-            _rlnMicrographShiftY #3
+loop_ 
+_rlnMicrographFrameNumber #1 
+_rlnMicrographShiftX #2 
+_rlnMicrographShiftY #3
 
 eot
 
-          awk '{ print NR, $1, $2 }' motioncor2_shifts.txt >> ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star
+          awk '{ print "         ", NR, "   ", $1, "   ", $2 }' motioncor2_shifts.txt >> ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star
 
       else
 
+        if ( -e motioncor2_shifts.star ) then
           cat motioncor2_shifts.star > ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star
 
           sed -i "/_rlnMicrographMovieName/c\_rlnMicrographMovieName                            ${export_rawstack_subdir}/${import_rawstack}"  ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star
 
           sed -i "/_rlnMicrographGainName/c\_rlnMicrographGainName                             ${export_gainref_subdir}/${gainref_name:r}.mrc"  ${sub_basedir}/${sub_targetdir}/${sub_targetname:r}.star
+          
+        else
+
+          echo "::WARNING: When working on "\"${sub_targetdir}\"", no drift-correction information was found."
+
+        endif
 
       endif
